git log --pretty=format:"%h %ad | %s%d [%an]" origin/abt/topic/gmin/l-dev/master..origin/abt/topic/gmin/l-dev/boot-devel/master:
  2099f1f Mon Jul 28 10:29:44 2014 -0700 | fix incorrect empty directories being made (origin/abt/topic/gmin/l-dev/boot-devel/master) [Andrew Boie]
  f543b76 Mon Jul 21 16:28:47 2014 -0700 | get rid of Android.mk [Andrew Boie]
  ffb02a4 Wed Jul 23 13:16:46 2014 -0700 | Allow override of default certs in shim [Matt Wood]
  b236ffb Mon Jul 21 16:18:56 2014 -0700 | generate-prebuilts.sh: stage built OpenSSL and headers to prebuilts [Andrew Boie]
  2fc30c4 Wed Jul 9 15:02:50 2014 +0800 | Update openssl to 0.9.8za [Gary Ching-Pang Lin]

git diff --stat origin/abt/topic/gmin/l-dev/master..origin/abt/topic/gmin/l-dev/boot-devel/master:
  Android.mk                                |  44 --
  Cryptlib/Cryptlib.diff                    |   4 +-
  Cryptlib/Include/openssl/bn.h             |  11 +
  Cryptlib/Include/openssl/crypto.h         |  37 +-
  Cryptlib/Include/openssl/ec.h             |  10 +-
  Cryptlib/Include/openssl/engine.h         |   8 +-
  Cryptlib/Include/openssl/opensslv.h       |   6 +-
  Cryptlib/Include/openssl/ssl.h            |  13 +-
  Cryptlib/Include/openssl/ssl3.h           |  10 +
  Cryptlib/Include/openssl/symhacks.h       |  10 +-
  Cryptlib/Include/openssl/tls1.h           |  14 +
  Cryptlib/Makefile                         |   6 +-
  Cryptlib/OpenSSL/crypto/asn1/a_int.c      |   2 +-
  Cryptlib/OpenSSL/crypto/asn1/a_strex.c    |   1 +
  Cryptlib/OpenSSL/crypto/asn1/a_strnid.c   |   2 +-
  Cryptlib/OpenSSL/crypto/asn1/a_verify.c   |   6 +
  Cryptlib/OpenSSL/crypto/asn1/t_pkey.c     |   5 -
  Cryptlib/OpenSSL/crypto/asn1/x_pubkey.c   |   5 +-
  Cryptlib/OpenSSL/crypto/bn/bn_lib.c       |  52 ++
  Cryptlib/OpenSSL/crypto/bn/bn_mont.c      |  46 +-
  Cryptlib/OpenSSL/crypto/bn/bn_word.c      |  25 +-
  Cryptlib/OpenSSL/crypto/cryptlib.c        |  16 +
  Cryptlib/OpenSSL/crypto/ec/ec2_mult.c     |  26 +-
  Cryptlib/OpenSSL/crypto/ec/ec_key.c       |  13 +-
  Cryptlib/OpenSSL/crypto/ec/ec_lib.c       |  10 +-
  Cryptlib/OpenSSL/crypto/ecdh/ech_lib.c    |  11 +-
  Cryptlib/OpenSSL/crypto/ecdsa/ecs_lib.c   |  11 +-
  Cryptlib/OpenSSL/crypto/engine/eng_all.c  |   6 +-
  Cryptlib/OpenSSL/crypto/err/err_all.c     |   2 +
  Cryptlib/OpenSSL/crypto/evp/bio_b64.c     |   3 +-
  Cryptlib/OpenSSL/crypto/evp/encode.c      |   1 +
  Cryptlib/OpenSSL/crypto/o_init.c          |  14 +
  Cryptlib/OpenSSL/crypto/ocsp/ocsp_vfy.c   |  10 +-
  Cryptlib/OpenSSL/crypto/pkcs12/p12_crt.c  |   8 +
  Cryptlib/OpenSSL/crypto/pkcs12/p12_kiss.c |   2 +-
  Cryptlib/OpenSSL/crypto/pkcs7/pk7_smime.c |   4 +-
  Cryptlib/OpenSSL/crypto/rsa/rsa_oaep.c    |   2 +-
  Cryptlib/OpenSSL/crypto/x509/x509_vfy.c   |   6 +-
  Cryptlib/OpenSSL/update.sh                | 998 +++++++++++++++---------------
  Cryptlib/Pk/CryptPkcs7.c                  | 753 ----------------------
  Cryptlib/Pk/CryptPkcs7Sign.c              | 207 +++++++
  Cryptlib/Pk/CryptPkcs7SignNull.c          |  59 ++
  Cryptlib/Pk/CryptPkcs7Verify.c            | 677 ++++++++++++++++++++
  Cryptlib/Pk/CryptPkcs7VerifyNull.c        | 100 +++
  Cryptlib/Pk/CryptRsaExtNull.c             | 125 ++++
  Cryptlib/Rand/CryptRand.c                 |  11 +-
  Cryptlib/update.sh                        |   9 +-
  generate-prebuilts.sh                     |  42 +-
  48 files changed, 2019 insertions(+), 1424 deletions(-)

+++ generate-prebuilts.sh
@@ -5,6 +5,12 @@
 # Please make sure you have Android's build system setup first, and lunch
 # target defined.
 #
+# Setting the SHIM_KEY_PAIR variable will cause the shim to be built with the
+# specified shim certificate instead of generating random certificate.
+#
+# Setting VENDOR_KEY_PAIR overrides the default test key to be embedded in the
+# shim. If set to empty string, no vendor cert will be embedded.
+#
 # Also, it requires the pesign utility from
 #     https://github.com/vathpela/pesign
 # You will need to build this by yourself using GCC 4.7.
@@ -26,17 +32,27 @@ if [ -z "$ANDROID_BUILD_TOP" ]; then
 fi
 
 DB_KEY_PAIR=$ANDROID_BUILD_TOP/device/intel/build/testkeys/DB
-VENDOR_KEY_PAIR=$ANDROID_BUILD_TOP/device/intel/build/testkeys/vendor
 
 rm -f db.key
 openssl pkcs8 -nocrypt -inform DER -outform PEM -in ${DB_KEY_PAIR}.pk8 -out db.key
 
-rm -f vendor.cer
-openssl x509 -outform der -in ${VENDOR_KEY_PAIR}.x509.pem -out vendor.cer
+if [ ! -v VENDOR_KEY_PAIR ]; then
+  VENDOR_KEY_PAIR=$ANDROID_BUILD_TOP/device/intel/build/testkeys/vendor
+fi
+if [ -z "$VENDOR_KEY_PAIR" ]; then
+  echo "Using empty vendor key pair"
+  VENDOR_CERT_PARAM=
+else
+  echo "Using vendor key pair $VENDOR_KEY_PAIR"
+  rm -f vendor.cer
+  openssl x509 -outform der -in ${VENDOR_KEY_PAIR}.x509.pem -out vendor.cer
+  VENDOR_CERT_PARAM="VENDOR_CERT_FILE=vendor.cer"
+fi
+
 
 copy_to_prebuilts()
 {
-    PREBUILT_FILES="MokManager.efi.signed shim.efi"
+    PREBUILT_FILES="MokManager.efi.signed shim.efi Cryptlib/OpenSSL/libopenssl.a"
 
     # Sanity check
     have_prebuilt_files=1
@@ -52,6 +68,7 @@ copy_to_prebuilts()
     fi
 
     # All files present. Copy them into prebuilts/
+    rm -rf $PREBUILT_TOP/uefi_shim/linux-$1/
     mkdir -p $PREBUILT_TOP/uefi_shim/linux-$1/
     cp -v MokManager.efi.signed $PREBUILT_TOP/uefi_shim/linux-$1/MokManager.efi
 
@@ -59,6 +76,8 @@ copy_to_prebuilts()
 	    --output shim.efi.signed shim.efi
 
     cp -v shim.efi.signed $PREBUILT_TOP/uefi_shim/linux-$1/shim.efi
+    cp -v Cryptlib/OpenSSL/libopenssl.a $PREBUILT_TOP/uefi_shim/linux-$1/libopenssl.a
+    cp -rf Cryptlib/Include $PREBUILT_TOP/uefi_shim/linux-$1/
 }
 
 add_prebuilts=0
@@ -93,10 +112,6 @@ if [ "$have_all_files" == "0" ]; then
     exit 1
 fi
 
-# Clean up everything and create prebuilts directory
-mkdir -p $PREBUILT_TOP/linux-x86/uefi_shim
-mkdir -p $PREBUILT_TOP/linux-x86_64/uefi_shim
-
 # Upstream Makefile heavily biased towards Fedora, we need to override
 # a bunch of Make variables
 EFI_PATH_64=$PREBUILT_TOP/gnu-efi/linux-x86_64
@@ -115,17 +130,24 @@ MAKE_CMD="make -j12 DEFAULT_LOADER=\\\\\\\\gummiboot.efi"
 $MAKE_CMD ARCH=x86_64 clean
 $MAKE_CMD ARCH=ia32 clean
 
+if [ -v SHIM_KEY_PAIR ]; then
+    echo "Using shim key $SHIM_KEY_PAIR"
+    cp $SHIM_KEY_PAIR.x509.pem shim.crt
+else
+    echo "Using default random cert in shim..."
+fi
+
 # Generate prebuilts for x86_64
 $MAKE_CMD ARCH=x86_64 EFI_PATH=$EFI_PATH_64 LIB_PATH=$LIB_PATH_64 \
           EFI_INCLUDE=$EFI_INCLUDE_64 EFI_CRT_OBJS=$EFI_CRT_OBJS_64 \
-          EFI_LDS=$EFI_LDS_64 VENDOR_CERT_FILE=vendor.cer
+          EFI_LDS=$EFI_LDS_64 $VENDOR_CERT_PARAM
 copy_to_prebuilts x86_64
 $MAKE_CMD ARCH=x86_64 clean
 
 # Generate prebuilts for ia32
 $MAKE_CMD ARCH=ia32 EFI_PATH=$EFI_PATH_32 LIB_PATH=$LIB_PATH_32 \
           EFI_INCLUDE=$EFI_INCLUDE_32 EFI_CRT_OBJS=$EFI_CRT_OBJS_32 \
-          EFI_LDS=$EFI_LDS_32 VENDOR_CERT_FILE=vendor.cer
+          EFI_LDS=$EFI_LDS_32 $VENDOR_CERT_PARAM
 copy_to_prebuilts x86
 $MAKE_CMD ARCH=ia32 clean
